+++ lib/diff.ml
@@ -1,44 +1,41 @@
 open Types
 open Stddiff
 
+type typ_exp = Tuple of tuple | Atomic of type_expr atomic_modification
+and tuple = (type_expr, typ_exp) List_.t
+
 type type_modification = {
   type_kind : (type_decl_kind, type_kind) maybe_changed;
   type_privacy : (Asttypes.private_flag, type_privacy) maybe_changed;
-  type_manifest : type_expr Stddiff.atomic_option;
-  type_params : (type_expr, type_param) Stddiff.list_;
+  type_manifest : (type_expr, typ_exp) Option_.t;
+  type_params : (type_expr list, (type_expr, typ_exp) List_.t) maybe_changed;
 }
 
 and type_kind =
-  | Record_tk of (label_declaration, label) map
-  | Variant_tk of (Types.constructor_declaration, cstr_args) map
+  | Record_tk of (label_declaration, label) Map_.t
+  | Variant_tk of (constructor_declaration, cstr_args) Map_.t
   | Atomic_tk of type_decl_kind atomic_modification
 
 and label = {
-  label_type : type_expr Stddiff.maybe_changed_atomic;
+  label_type : (type_expr, typ_exp) maybe_changed;
   label_mutable : (Asttypes.mutable_flag, field_mutability) maybe_changed;
 }
 
 and field_mutability = Added_m | Removed_m
 
 and cstr_args =
-  | Record_cstr of (label_declaration, label) map
-  | Tuple_cstr of type_expr maybe_changed_atomic_entry list
-  | Atomic_cstr of Types.constructor_arguments Stddiff.atomic_modification
+  | Record_cstr of (label_declaration, label) Map_.t
+  | Tuple_cstr of tuple
+  | Atomic_cstr of constructor_arguments Stddiff.atomic_modification
 
 and type_privacy = Added_p | Removed_p
-and type_param = (type_expr, type_param_diff) maybe_changed
-and type_param_diff = Added_tp of type_expr | Removed_tp of type_expr
 
 type type_ = {
   tname : string;
   tdiff : (type_declaration, type_modification) entry;
 }
 
-type value = {
-  vname : string;
-  vdiff : (value_description, type_expr atomic_modification) entry;
-}
-
+type value = { vname : string; vdiff : (value_description, typ_exp) entry }
 type class_ = { cname : string; cdiff : class_declaration Stddiff.atomic_entry }
 
 type cltype = {
@@ -123,22 +120,39 @@ let module_type_fallback ~loc ~typing_env ~name ~reference ~current =
 let expand_alias_types ~typing_env ~type_expr =
   Ctype.full_expand ~may_forget_scope:false typing_env type_expr
 
-let type_expr ~typing_env ?(ref_params = []) ?(cur_params = []) reference
+let rec type_expr ~typing_env ?(ref_params = []) ?(cur_params = []) reference
     current =
-  let normed_ref, normed_cur =
-    Normalize.type_params_arity ~reference:ref_params ~current:cur_params
-  in
-  if
-    Ctype.is_equal typing_env true
-      (normed_ref @ [ reference ])
-      (normed_cur @ [ current ])
-  then None
-  else
-    Some
-      {
-        reference = expand_alias_types ~typing_env ~type_expr:reference;
-        current = expand_alias_types ~typing_env ~type_expr:current;
-      }
+  match (Types.get_desc reference, Types.get_desc current) with
+  | Ttuple ref_exps, Ttuple cur_exps -> (
+      let type_exprs =
+        type_exprs ~typing_env ~ref_params ~cur_params ~reference:ref_exps
+          ~current:cur_exps
+      in
+      match type_exprs with
+      | Same _ -> Same reference
+      | Changed change -> Changed (Tuple change))
+  | _ ->
+      let normed_ref, normed_cur =
+        Normalize.type_params_arity ~reference:ref_params ~current:cur_params
+      in
+      if
+        Ctype.is_equal typing_env true
+          (normed_ref @ [ reference ])
+          (normed_cur @ [ current ])
+      then Same reference
+      else
+        Changed
+          (Atomic
+             {
+               reference = expand_alias_types ~typing_env ~type_expr:reference;
+               current = expand_alias_types ~typing_env ~type_expr:current;
+             })
+
+and type_exprs ~typing_env ~ref_params ~cur_params ~reference ~current =
+  List_.diff
+    ~diff_one:(fun ref cur ->
+      type_expr ~typing_env ~ref_params ~cur_params ref cur)
+    ~reference ~current
 
 let rec type_item ~typing_env ~name ~reference ~current =
   match (reference, current) with
@@ -187,7 +201,7 @@ and type_kind ~typing_env ~ref_params ~cur_params ~reference ~current =
         record_type ~typing_env ~ref_params ~cur_params ~ref_label_lst
           ~cur_label_lst
       in
-      if String_map.is_empty label_map.changed_map then Same reference
+      if String_map.is_empty label_map.Map_.changed_map then Same reference
       else Changed (Record_tk label_map)
   | Type_variant (ref_constructor_lst, _), Type_variant (cur_constructor_lst, _)
     ->
@@ -195,7 +209,7 @@ and type_kind ~typing_env ~ref_params ~cur_params ~reference ~current =
         variant_type ~typing_env ~ref_params ~cur_params ~ref_constructor_lst
           ~cur_constructor_lst
       in
-      if String_map.is_empty cstr_map.changed_map then Same reference
+      if String_map.is_empty cstr_map.Map_.changed_map then Same reference
       else Changed (Variant_tk cstr_map)
   | Type_abstract _, Type_abstract _ -> Same reference
   | Type_open, Type_open -> Same reference
@@ -206,9 +220,9 @@ and record_type ~typing_env ~ref_params ~cur_params ~ref_label_lst
     ~cur_label_lst =
   let ref_lbls = extract_lbls ref_label_lst in
   let cur_lbls = extract_lbls cur_label_lst in
-  diff_map
+  Map_.diff
     ~diff_one:(label ~typing_env ~ref_params ~cur_params)
-    ~ref_map:ref_lbls ~cur_map:cur_lbls
+    ~reference:ref_lbls ~current:cur_lbls
 
 and label ~typing_env ~ref_params ~cur_params reference current =
   let label_type =
@@ -219,11 +233,11 @@ and label ~typing_env ~ref_params ~cur_params reference current =
     label_mutable ~reference:reference.ld_mutable ~current:current.ld_mutable
   in
   match (label_type, label_mutable) with
-  | None, Same _ -> None
-  | None, label_mutable ->
-      Some { label_type = Same reference.ld_type; label_mutable }
-  | Some type_diff, label_mutable ->
-      Some { label_type = Changed type_diff; label_mutable }
+  | Same _, Same _ -> Same reference
+  | Same _, label_mutable ->
+      Changed { label_type = Same reference.ld_type; label_mutable }
+  | Changed type_diff, label_mutable ->
+      Changed { label_type = Changed type_diff; label_mutable }
 
 and label_mutable ~reference ~current =
   match (reference, current) with
@@ -237,52 +251,34 @@ and variant_type ~typing_env ~ref_params ~cur_params ~ref_constructor_lst
     ~cur_constructor_lst =
   let ref_cstrs = extract_cstrs ref_constructor_lst in
   let cur_cstrs = extract_cstrs cur_constructor_lst in
-  diff_map
+  Map_.diff
     ~diff_one:(cstr ~typing_env ~ref_params ~cur_params)
-    ~ref_map:ref_cstrs ~cur_map:cur_cstrs
+    ~reference:ref_cstrs ~current:cur_cstrs
 
 and cstr ~typing_env ~ref_params ~cur_params reference current =
   match (reference.cd_args, current.cd_args) with
-  | Cstr_tuple ref_tuple, Cstr_tuple cur_tuple -> (
-      let tuple =
-        tuple_type ~typing_env ~ref_params ~cur_params ~reference:ref_tuple
-          ~current:cur_tuple
+  | Cstr_tuple ref_type_exprs, Cstr_tuple cur_type_exprs -> (
+      let type_exprs =
+        type_exprs ~typing_env ~ref_params ~cur_params ~reference:ref_type_exprs
+          ~current:cur_type_exprs
       in
-      match tuple with Same _ -> None | Changed diff -> Some (Tuple_cstr diff))
+      match type_exprs with
+      | Same _ -> Same reference
+      | Changed change -> Changed (Tuple_cstr change))
   | Cstr_record ref_record, Cstr_record cur_record ->
       let label_map =
         record_type ~typing_env ~ref_params ~cur_params
           ~ref_label_lst:ref_record ~cur_label_lst:cur_record
       in
-      if String_map.is_empty label_map.changed_map then None
-      else Some (Record_cstr label_map)
+      if String_map.is_empty label_map.changed_map then Same reference
+      else Changed (Record_cstr label_map)
   | _ ->
-      Some
+      Changed
         (Atomic_cstr
            { reference = reference.cd_args; current = current.cd_args })
 
-and tuple_type ~typing_env ~ref_params ~cur_params ~reference ~current =
-  diff_list
-    ~diff_one:(fun t1 t2 ->
-      match (t1, t2) with
-      | None, None -> assert false
-      | Some t1, None -> Changed (Removed t1)
-      | None, Some t2 -> Changed (Added t2)
-      | Some t1, Some t2 -> (
-          match type_expr ~typing_env ~ref_params ~cur_params t1 t2 with
-          | None -> Same t1
-          | Some diff -> Changed (Modified diff)))
-    ~ref_list:reference ~cur_list:current
-
 and type_params ~reference ~current =
-  diff_list
-    ~diff_one:(fun t1 t2 ->
-      match (t1, t2) with
-      | None, None -> assert false
-      | Some t1, None -> Changed (Removed_tp t1)
-      | None, Some t2 -> Changed (Added_tp t2)
-      | Some t1, Some _ -> Same t1)
-    ~ref_list:reference ~cur_list:current
+  List_.diff ~diff_one:(fun t1 _ -> Same t1) ~reference ~current
 
 and type_privacy ~reference ~current =
   match (reference, current) with
@@ -292,14 +288,9 @@ and type_privacy ~reference ~current =
   | Asttypes.Private, Asttypes.Private -> Same Asttypes.Private
 
 and type_manifest ~typing_env ~ref_params ~cur_params ~reference ~current =
-  match (reference, current) with
-  | None, None -> Same None
-  | Some t1, None -> Changed (Removed t1)
-  | None, Some t2 -> Changed (Added t2)
-  | Some t1, Some t2 -> (
-      match type_expr ~typing_env ~ref_params ~cur_params t1 t2 with
-      | None -> Same (Some t1)
-      | Some diff -> Changed (Modified diff))
+  Option_.diff
+    ~diff_one:(type_expr ~typing_env ~ref_params ~cur_params)
+    ~reference ~current
 
 let value_descripiton ~typing_env reference current =
   type_expr ~typing_env reference.val_type current.val_type
@@ -313,8 +304,8 @@ let value_item ~typing_env ~name ~reference ~current =
   | Some reference, Some current -> (
       let val_type_diff = value_descripiton ~typing_env reference current in
       match val_type_diff with
-      | None -> None
-      | Some type_expr_diff ->
+      | Same _ -> None
+      | Changed type_expr_diff ->
           Some (Value { vname = name; vdiff = Modified type_expr_diff }))
 
 let class_item ~typing_env ~name ~(reference : class_declaration option)