+++ lib/text_diff.ml
@@ -240,7 +240,6 @@ and process_privacy_diff privacy_diff =
 
 and process_type_params_diff params_diff =
   let open Stddiff in
-  let open Diff in
   let params_hunks =
     match params_diff with
     | Same params ->
@@ -257,12 +256,13 @@ and process_type_params_diff params_diff =
             | Same same_param ->
                 Icommon
                   (Printf.sprintf "%s%s" comma (type_expr_to_string same_param))
-            | Changed (Added_tp p) ->
+            | Changed (Added p) ->
                 Iconflict
                   { iorig = None; inew = Some (comma ^ type_expr_to_string p) }
-            | Changed (Removed_tp p) ->
+            | Changed (Removed p) ->
                 Iconflict
-                  { iorig = Some (comma ^ type_expr_to_string p); inew = None })
+                  { iorig = Some (comma ^ type_expr_to_string p); inew = None }
+            | Changed (Modified _) -> assert false)
           changed_params
   in
   let open_paren = Icommon " (" in
@@ -346,15 +346,7 @@ and process_manifest_diff manifest_diff =
       [ Iconflict { iorig = None; inew = Some (" " ^ type_expr_to_string te) } ]
   | Changed (Removed te) ->
       [ Iconflict { iorig = Some (" " ^ type_expr_to_string te); inew = None } ]
-  | Changed (Modified { reference; current }) ->
-      [
-        Icommon " ";
-        Iconflict
-          {
-            iorig = Some (type_expr_to_string reference);
-            inew = Some (type_expr_to_string current);
-          };
-      ]
+  | Changed (Modified te_diff) -> Icommon " " :: process_type_expr_diff te_diff
 
 and process_type_kind_diff kind_diff =
   match kind_diff with
@@ -368,7 +360,7 @@ and process_type_kind_diff kind_diff =
       [ Line_conflict { orig; new_ } ]
 
 and process_record_type_diff record_diff =
-  let open Stddiff in
+  let open Stddiff.Map_ in
   let { same_map = same_lbls; changed_map = changed_lbls } = record_diff in
   let common_hunks =
     List.map
@@ -399,15 +391,7 @@ and process_label_diff name label_diff =
 and process_label_type_diff label_type_diff =
   match label_type_diff with
   | Stddiff.Same te -> [ Icommon (" " ^ type_expr_to_string te) ]
-  | Stddiff.Changed { reference; current } ->
-      [
-        Icommon " ";
-        Iconflict
-          {
-            iorig = Some (type_expr_to_string reference);
-            inew = Some (type_expr_to_string current);
-          };
-      ]
+  | Stddiff.Changed te_diff -> Icommon " " :: process_type_expr_diff te_diff
 
 and process_mutablity_diff mutablity_diff =
   let open Stddiff in
@@ -426,7 +410,7 @@ and type_kind_to_lines type_kind =
   | Type_open -> [ ".." ]
 
 and process_variant_type_diff variant_diff =
-  let open Stddiff in
+  let open Stddiff.Map_ in
   let { same_map = same_cstrs; changed_map = changed_cstrs } = variant_diff in
   let common_hunks =
     List.map
@@ -482,17 +466,24 @@ and process_tuple_type_diff tuple_diff =
             Iconflict
               { iorig = Some (star ^ type_expr_to_string te); inew = None };
           ]
-      | Changed (Modified { reference; current }) ->
-          let te_hunk =
-            Iconflict
-              {
-                iorig = Some (type_expr_to_string reference);
-                inew = Some (type_expr_to_string current);
-              }
-          in
-          if i > 0 then [ Icommon " * "; te_hunk ] else [ te_hunk ])
+      | Changed (Modified te) ->
+          let te_hunks = process_type_expr_diff te in
+          if i > 0 then Icommon " * " :: te_hunks else te_hunks)
     tuple_diff
   |> List.concat
+
+and process_type_expr_diff (diff : Diff.typ_exp) : inline_hunk list =
+  match diff with
+  | Diff.Atomic { reference; current } ->
+      [
+        Iconflict
+          {
+            iorig = Some (type_expr_to_string reference);
+            inew = Some (type_expr_to_string current);
+          };
+      ]
+  | Tuple tuple_diff ->
+      (Icommon "(" :: process_tuple_type_diff tuple_diff) @ [ Icommon ")" ]
 
 and cstr_args_to_line cstr_args =
   match cstr_args with
@@ -514,15 +505,10 @@ and lbls_to_line lbls =
 and tuple_to_line tuple =
   List.map type_expr_to_string tuple |> String.concat " * "
 
-let process_vd_diff name
-    ({ reference; current } : Types.type_expr Stddiff.atomic_modification) =
+let process_vd_diff name diff =
   let header = Icommon (Format.sprintf "val %s : " name) in
-  let type_ =
-    let iorig = Some (type_expr_to_string reference) in
-    let inew = Some (type_expr_to_string current) in
-    Iconflict { iorig; inew }
-  in
-  [ Inline_hunks [ header; type_ ] ]
+  let type_ = process_type_expr_diff diff in
+  [ Inline_hunks (header :: type_) ]
 
 let process_value_diff (val_diff : Diff.value) =
   process_entry ~entry_to_string:vd_to_lines